Added theme support. Changed file structure. Added missing license information for Lato font-files.

+++ clear_cache.php
@@ -2,17 +2,15 @@
 <html lang="de" dir="ltr">
 <head>
   <meta charset="utf-8">
-  <link  rel="stylesheet" type="text/css" href="http://fonts.googleapis.com/css?family=Lato:100,400">
-  <link  rel="stylesheet" type="text/css" href="res/css/default.css" media="screen">
   <title>Clear Cache</title>
 </head>
 <body>
-<div class="page-wrapper"><?php
-  require('inc/config.php');
+<?php
+  require('includes/config.php');
 
   if (CLEAR_THUMB_SECRET == '' || (isset($_GET['s']) && CLEAR_THUMB_SECRET == $_GET['s'])) {
     function recursiveDelete($str) {
-        if (is_file($str)) {
+        if (is_file($str) && basename($str) != 'index.html') {
           echo '<li>File: '.$str.'</li>';
           return @unlink($str);
         } else if (is_dir($str)) {
@@ -32,6 +30,5 @@
     echo '</ol></p>...Done</p>';
   }
 ?>
-</div>
 </body>
 </html>
